M_15_queries/requirements.txt
asgiref==3.5.2
Django==4.1.6
django-filter==22.1
django-taggit==4.0.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.2
